P-40 Make error hsndling while updating posts and comments

+++ backend/services/GatherService/Services/InfoService/IInfoService.cs
@@ -26,7 +26,7 @@ public interface IInfoService
 
     Task UpdatePostComments(long chatId, long postId, WebSocket webSocket);
 
-    Task<ServiceResponse<int>> GetCommentsCount(long chatId, long postId);
+    Task<ServiceResponse<long>> GetCommentsCount(long chatId, long postId);
 
     Task<ServiceResponse<IEnumerable<CommentDto>>> GetComments(long chatId, long postId, int offset = 0, int limit = 0);
